Fix file name and usage after class rename
