+++ example/lib/ui/push_notifications/push_notifications_device_information.dart
@@ -1,6 +1,6 @@
 import 'package:ably_flutter/ably_flutter.dart' as ably;
 import 'package:ably_flutter_example/push_notifications/push_notification_service.dart';
-import 'package:ably_flutter_example/ui/text_with_label.dart';
+import 'package:ably_flutter_example/ui/text_row.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
