+++ example/lib/ui/push_notifications/push_notifications_ios_permissions_sliver.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:ably_flutter/ably_flutter.dart' as ably;
 import 'package:ably_flutter_example/push_notifications/push_notification_service.dart';
 import 'package:ably_flutter_example/ui/bool_stream_button.dart';
-import 'package:ably_flutter_example/ui/text_with_label.dart';
+import 'package:ably_flutter_example/ui/text_row.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
